move details to body container
